requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.2.0
bidict==0.23.1
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
h11==0.14.0
idna==3.7
MouseInfo==0.1.3
multidict==6.0.5
netifaces==0.10.6
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
python-engineio==4.9.1
python-socketio==5.11.3
pytweening==1.2.0
requests==2.32.3
rubicon-objc==0.4.9
setuptools==3.3
simple-websocket==1.0.0
urllib3==2.2.2
wsproto==1.2.0
yarl==1.9.4
